feat: Consolidate AMD GSM8K evaluation tests into a single script
- Migrated 1-GPU and 2-GPU evaluation tests into a unified test script for AMD models.
- Removed outdated individual test files for 1-GPU and 2-GPU evaluations.
- Updated model score thresholds and failing models list for improved accuracy in evaluations.
- Enhanced server launch configurations and reporting for better clarity in test results.
